+++ lib/couchdb.js
@@ -9,50 +9,9 @@ var url = require('url'),
     http = require('http'),
     https = require('https'),
     logger = require('./logger'),
+    utils = require('./utils'),
     querystring = require('querystring'),
     _ = require('underscore');
-
-var STATUS_MSGS = {
-    400: '400: Bad Request',
-    401: '401: Unauthorized',
-    402: '402: Payment Required',
-    403: '403: Forbidden',
-    404: '404: Not Found',
-    405: '405: Method Not Allowed',
-    406: '406: Not Acceptable',
-    407: '407: Proxy Authentication Required',
-    408: '408: Request Timeout',
-    409: '409: Conflict',
-    410: '410: Gone',
-    411: '411: Length Required',
-    412: '412: Precondition Failed',
-    413: '413: Request Entity Too Large',
-    414: '414: Request-URI Too Long',
-    415: '415: Unsupported Media Type',
-    416: '416: Requested Range Not Satisfiable',
-    417: '417: Expectation Failed',
-    418: '418: I\'m a teapot',
-    422: '422: Unprocessable Entity',
-    423: '423: Locked',
-    424: '424: Failed Dependency',
-    425: '425: Unordered Collection',
-    444: '444: No Response',
-    426: '426: Upgrade Required',
-    449: '449: Retry With',
-    450: '450: Blocked by Windows Parental Controls',
-    499: '499: Client Closed Request',
-    500: '500: Internal Server Error',
-    501: '501: Not Implemented',
-    502: '502: Bad Gateway',
-    503: '503: Service Unavailable',
-    504: '504: Gateway Timeout',
-    505: '505: HTTP Version Not Supported',
-    506: '506: Variant Also Negotiates',
-    507: '507: Insufficient Storage',
-    509: '509: Bandwidth Limit Exceeded',
-    510: '510: Not Extended'
-};
-
 
 var CouchDB;
 
@@ -219,7 +178,7 @@ CouchDB.prototype.client = function (method, path, data, callback) {
         //});
         if (response.headers.connection === 'close' &&
             response.statusCode >= 300) {
-            var err3 = exports.statusCodeError(response.statusCode);
+            var err3 = utils.statusCodeError(response.statusCode);
             err3.response = response;
             callback(err3, data, response);
         }
@@ -247,7 +206,7 @@ CouchDB.prototype.client = function (method, path, data, callback) {
                         callback(err, data, response);
                     }
                     else {
-                        var err2 = exports.statusCodeError(response.statusCode);
+                        var err2 = utils.statusCodeError(response.statusCode);
                         err2.response = response;
                         callback(err2, data, response);
                     }
@@ -269,17 +228,6 @@ CouchDB.prototype.client = function (method, path, data, callback) {
     //logger.debug('request:', request.output[0]);
 };
 
-/**
- * Creates an error object with a message depending on the HTTP status code
- * of a response.
- */
-
-exports.statusCodeError = function (code) {
-    if (code in STATUS_MSGS) {
-        return new Error(STATUS_MSGS[code]);
-    }
-    return new Error('Status code: ' + code);
-};
 
 /**
  * Encodes a document id or view, list or show name.
