+++ lib/repository.js
@@ -444,14 +444,14 @@ exports.availableVersions = function (name, repositories, callback) {
             }
             if (doc && doc.versions) {
                 for (var k in doc.versions) if (doc.versions.hasOwnProperty(k)) {
-                    //if (!versions[k]) {
+                    if (!versions[k]) {
                         versions.push({
                             source: 'repository',
                             repository: repo,
                             config: doc.versions[k],
                             version: k
                         });
-                    //}
+                    }
                 }
             }
             cb();
