+++ lib/tree.js
@@ -186,7 +186,6 @@ exports.addDependency = function (parent, name, range, sources, packages, cb) {
                 .value();
 
             if (!_.isUndefined(matchHighestPriority)) {
-
                 prioritizedVersions = _.chain(dep.versions).where({ priority: matchHighestPriority }).pluck('version').value();
 
                 if (!_.isEmpty(prioritizedVersions)) {
