+++ lib/utils.js
@@ -10,6 +10,46 @@ var fs = require('fs'),
     urlParse = url.parse,
     _ = require('underscore');
 
+var STATUS_MSGS = {
+    400: '400: Bad Request',
+    401: '401: Unauthorized',
+    402: '402: Payment Required',
+    403: '403: Forbidden',
+    404: '404: Not Found',
+    405: '405: Method Not Allowed',
+    406: '406: Not Acceptable',
+    407: '407: Proxy Authentication Required',
+    408: '408: Request Timeout',
+    409: '409: Conflict',
+    410: '410: Gone',
+    411: '411: Length Required',
+    412: '412: Precondition Failed',
+    413: '413: Request Entity Too Large',
+    414: '414: Request-URI Too Long',
+    415: '415: Unsupported Media Type',
+    416: '416: Requested Range Not Satisfiable',
+    417: '417: Expectation Failed',
+    418: '418: I\'m a teapot',
+    422: '422: Unprocessable Entity',
+    423: '423: Locked',
+    424: '424: Failed Dependency',
+    425: '425: Unordered Collection',
+    444: '444: No Response',
+    426: '426: Upgrade Required',
+    449: '449: Retry With',
+    450: '450: Blocked by Windows Parental Controls',
+    499: '499: Client Closed Request',
+    500: '500: Internal Server Error',
+    501: '501: Not Implemented',
+    502: '502: Bad Gateway',
+    503: '503: Service Unavailable',
+    504: '504: Gateway Timeout',
+    505: '505: HTTP Version Not Supported',
+    506: '506: Variant Also Negotiates',
+    507: '507: Insufficient Storage',
+    509: '509: Bandwidth Limit Exceeded',
+    510: '510: Not Extended'
+};
 
 var pathExists = fs.exists || path.exists;
 
@@ -239,6 +279,17 @@ exports.truncate = function (str, max) {
     return str.substr(0, max - 1) + '…';
 };
 
+/**
+ * Creates an error object with a message depending on the HTTP status code
+ * of a response.
+ */
+
+exports.statusCodeError = function (code) {
+    if (code in STATUS_MSGS) {
+        return new Error(STATUS_MSGS[code]);
+    }
+    return new Error('Status code: ' + code);
+};
 
 exports.download = function (file, target, callback) {
     var urlinfo = url.parse(file);
@@ -281,7 +332,7 @@ exports.download = function (file, target, callback) {
         req.on('response', function (response) {
             if (response.statusCode >= 300) {
                 this.abort();
-                return callback(couchdb.statusCodeError(response.statusCode));
+                return callback(exports.statusCodeError(response.statusCode));
             }
         }).on('error', function (err) {
             callback(err);
