+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jamjs",
-  "version": "0.3.2-rc.0",
+  "version": "0.3.3-rc.0",
   "description": "",
   "maintainers": [
     {
